Finished average RGB value function getRegionPixelAverage.
The method getRegionPixelAverage in the Region class calls the method
getRegionPixelList to gather a list of the colors of all the pixels
within that region, find a sum for each value among the R, G, and B values,
and divide those sums by the number of colors to find the average of R, G, and B
individually.

I want to work on finding the "dominance" of a color to see if a region is
overall red, green, or blue.