New RegionChunk class; new global "check" functions; more work on Region
Global function imageBoundsCheck returns True if a coordinate (x, y) is
within the bounds of the image; imageBoundsCheckRadius is very similar
but takes a radius into account: the radius cannot extend out of the image.

In the RegionChunk class, method chunkDefineExact creates a grid of Region
objects by starting from a specific location/corner and creating a given
number of Regions that are some pixels apart based on the "jump" parameters.

In the Region class, it is now possible to fill in a Region object of a
specific color with imageFillRegion.

New test functions regionTestA and chunkTestA test the Region and RegionChunk
class respectively by creating an object of each class and interacting with
some of the important methods that have been written.