Began work on background/red detection with code and planning docs
In the planning/ directory, a new file /redBackgroundCheckRules.md
gives a short summary on the approach to detecting background/red parts
of the image.

Global function "redBackgroundCheck" was added to show the "minimum" red
value for a corresponding color to see if that color would be considered
"red" and thus a background.

In RegionChunkm chunkDefineExact was modified to have 2D int tuple
parameters instead of individual X and Y coordinate parameters for
the starting position, jump lengths, etc.

Also, chunkDefinePack was created as an easier version of
chunkDefineExact where only a top-left corner and a bottom-right corner
coordinate (2D int tuple) needed to be specified with a radius in order
to define a grid of Region objects (much less manual and higher level).

Test functions chunkTestB and chunkTestC were created to test these
new features.