New background detection test code; encountered issues for redBackgroundCheck
The new test code is in the global function redBackgroundCheckDebugSimulation
where a completely separate, blank image is created and filled with a palette
of different shades of red, representing the "minimum" red colors for corresponding
values of Greeen and Blue (Red is the dependent variable/value, and Green and Blue would
both be the same independent variable/value).

POTENTIAL BUG IN Region's getRegionPixelList: in the past, a +1 was added to the rightmost
(horizontal), and bottommost (vertical) bounds so that an expected, accurate number
of pixels was generated in the list; however, this now results in an index
out of bounds error when the image has a resolution IDENTICAL to the size
of a region, so it has been removed.