Red detection has now been finalized across entirity of processor.py
Function redBackgroundCheck is now the universal function for testing
if a color is considered a part of the background/red. Therefore
the method getColorDominance in Region was removed.

Method isRegionRed in Region calls redBackgroundCheck to check if a
Region object is "red" in terms of its pixels.

Method chunkLabelRedDetectRegions in RegionChunk then calls isRegionRed
to fill in all regions that are considered background/red with the
color pink for debug purposes.