Additional chunk definition function added to RegionChunk
In chunkDefinePack, when determing the radius by diving the height and width
by the given parameter of the number of regions, it will be default use the
minimum between those two divisions, but now the maximum can be chosen with useMax.

chunkDefinePackAuto was written: instead of giving a user-defined, fixed number
of regions like in chunkDefinePack, now a radius can be given and the number
of regions is calculated automatically. This is because having a fixed region radius
now only produces chunks with more consistent shapes, BUT ALSO it allows you
to generally adjust the precision and coarseness with the region radius.