RegionChunk can now detect background/red across its own regions
The method chunkRedDetectRegions in RegionChunk can now detect Regions that
are considered as background/red. There are additional parameters that can
adjust how this is done: outlierThresh, filterBoundOutliers, and strictFilter.
Parameter filterBoundOutliers determines if outliers should be filtered
whatsoever, and if so, outlierThresh and strictFilter have an effect.
An outlier is a background/red Region that is far away from the other
bodies and areas where many other background/red Regions are (thus should
not be considered as that will skew the cropping). Parameter outlierThresh
is the maximum number of pixels of distance between a Region and the average
background/red Region bounds or the boudns of the previous consecutive
background/red Region. If True, strictFilter forces the threshold to use
the average bounds only, not both average and consecutive.

chunkRedDetectRegions returns a dictionary with the list of regions that
were red, and the sum, average, and farthest bounds (top, bottom, left
and right) of each of the background/red regions. This is useful to determine
where to actually crop.

RegionChunk's chunkFillColor was also created to fill in all the regions
of the chunk with a certain color.

In Region, new class attribute self.pixBounds holds the "bounds" of the Region
(top, bottom, left, and right bounds): these bounds have the least (top) and
greatest (bottom) y-coordinate of the Region position, and the least (left)
and greatest (right) x-coordinate of hte Region position.

New global test function chunkTestB2: has a rough CLI interface for running
specific commands and tests. The new "rd" command has a flexible "attribute"
system for recording values of specific types from the user. This should be
expanded upon in the future to be used across all test functions.