Verbose debug in chunkRedDetectRegions; test functions chunkTestB2 and chunkTestB3
In class RegionChunk, method chunkRedDetectRegions, the parameter verboseDebug
toggles the printing of debug information and ALSO toggles artificial slowdowns
of the detection process to allow for me to read the print statements as the
detection is running.

Minor changes to chunkTestB2: now accounts for verboseDebug in chunkRedDetectRegions.

New function chunkTestB3: implements a rough, prototype-like code structure
for defining a top, bottom, and left chunk (right missing) around the textbook in the
sample images; uses chunkRedDetectRegions in each chunk to determine the
area of the image to crop; actually crops the image and shows the image in a window.
Command "rd" does the cropping process and "fill" fills the chunks with different colors.