Orientation detection implemented in Cropper class
The method orientationDetect in class Cropper determines if there are more
background/red regions on the left chunk or the right chunk: if the left
chunk has more red regions, then return "left," if the right chunk has more
red regions, then return "right;" the non-dominating chunk will NOT be used
for the cropping process (e.g. if left had less, then left chunk is ignored).

In method crop, the orientation detection was implemented. There may be some
red-regions on a side of the picture even if the red-shirt is not there
(meaning you have noise: parts of the image with unintentional or unwanted
red detections). For the images I am using, you are either seeing the left OR
right page of the book. You should only check the left side chunk if on a left
sided page, and the right side chunk on a right sided page. So, it detects
whether or not the page/image is a left or right sided page.

(Rough Commit Message)