Separated the all-in-one processor.py into three different project files
The processor.py file used to be the ONLY project file and contained
EVERYTHING: the final code itself, testing code, etc. To make the
project much more portable and to strip it of the very rough, dirty
testing code that was used to test the behavior of the project, processor.py
is now primary.py, testing.py, and interface.py, where interface.py is
currently empty and will contain code to allow for shell/CLI control over
the project.

In testing.py, there are currently 7 test functions that were selected and
called by modifying the source code directly: now when running testing.py,
a pretty UI allows you to select, execute, and interact with test functions
without modifying or (hopefully) looking at the source code in main().

NOTE: testing.py MAY/WILL break as primary.py changes over time. It will
NOT be maintained/updated to reflect changes. A new module or interface.py
should be written to feature handy debug and testing code for developers
that is much more clean, robust, and sustainable.

Additional command listings and help information has been added to test
functions chunkTestB2 and chunkTestB3. Function regionTestA was slightly
modified to work with the current version of primary.py.