Prettification: big comment banners, docstrings for major functions written.
Big comments banners indicating the separate areas of code for the classes
Region, Cropper, and RegionChunk, as well as the imports and global items have
been written.

Docstrings for major, crucial functions and methods that are likely to be used
by me or other developers in the future have been written. They generally
follow PEP 257 guidelines, but may not be perfect in line wrapping or clarity.
Returned values are in the summary of the docstring and parameters are
either not discussed, discussed conversationally, or bulleted out.