interface.py minimally functional; missing help docs, unit testing, etc.
In interface.py, the main function parsing the arguments and making relevant
calls to primary.py is commandParse. This function is extremely large and
should be broken into other global variables and functions.

In the function, argparse is used to structure the expected arguments of the
program and also provide an automated help menu system. The help summaries
for each individual argument has NOT been finished yet (all help-related
code is VERY, VERY ROUGH: I lazily defined a nested function helpMsgCycle
within commandParse that cycles through the help strings in the list helpMsg
so I do not have to reference each list item by an index number (fixed index
numbers will be used in the future).

Library pathlib is also used to handle paths to image files: I am using
POSIX Paths, NOT regular Path objects; thus, interface.py MAY/WILL break
when used in Windows (as there are Windows Paths).

In terms of cropping, there is either the "book" or "box" approach: the
"book" method is optimized for the sample pictures in the repo, and "box"
is much more generic and RECOMMENDED for any other image.

Overall, the system IS FUNCTIONAL and CAN BE RAN (tested with --book so far).