Implemented -y option for filling in background/red recognized regions.
Now when the user toggles -y, all regions that are considered a part of the
background or "red" around the subject of a given image will be filled with
pink; all other regions are left unaffected. The written image file is
NOT cropped, otherwise the background/red regions will literally be
cropped off and the point of this debug flag is lost.

Since only one debug option should be used at a time, an error will appear
if the user uses -x and -y simultaneously. Also, off-topic from debugging,
the user must have the outlier filter enabled (-f) in order to use any
other filter related arguments.

In order for this debug flag to have real functionality, interface.py
now calls the brand new Cropper chunkFillRedDetect method: this
method takes in almost identical parameters as the RegionChunk
chunkRedDetectRegions method.