Completely finished help messages; more comments; extra checks and functionality.
helpMsg was changed from a list to a dict: this is so that string-based
keys can be used to reference help entries instead of integer-based keys
(so in the future new arguments are added or arguments are changed, it will
be much less of a hassle).

To reflect these changes, all of the argparse add_argument calls have been
updated to reference helpMsg as a dict. Also, for all single-value arguments,
nargs was changed from '?' to 1 because those values are MANDATORY; '?' implies
that the value is optional and the argument can be used without a value
(in this case, this SHOULD NOT be done). Furthermore, single-value arguments
now have their default values as lists, AND (regardless if the user provided
input or the default was used) the single-value args are converted from a
list to single values. This is because, even for single-value args, a list
is always returned (just like nargs = 2, 3, 4...).

Originally I intended the default inner box to literally be at the extreme
center of an image; upon testing, this proved extremely impractical and
results in images that are a few pixels wide (NOT GOOD). The default now is
that the inner box is defined using the sides of the outer box but offsetted
by some number of pixels (that number of pixels is currently 300).

Junk comments and (relatively) better comments have been written. I don't think
any extensive work for comments should be done for now.

BUG FIX IN primary.py: chunkSetup accidentally used the upper bound for the
inner box instead of the left bound when defining the left chunk.

EXISTING BUGS:
1) When an insanely large, positive crop margin is given, there are large
black regions of the final image (as if the image was somehow "anti-cropped").
There needs to be changes in primary.py to account for this.
2) Image file names with many extensions or periods WILL break interface.py
as the program only accounts for one extension.
3) We need to watch out for NEGATIVE VALUES and overall, beyond cropping,
pre-conditions need to be defined and filtered for so no crazy values can
sneak past.